requirements.txt
streamlit
python-dotenv

langchain
langchain-openai
langgraph

transformers

pinecone
langchain-pinecone
pypdf

mcp
langchain-mcp-adapters

tiktoken
pandas
numpy

langchain-community
langchain-text-splitters
sentence-transformers
torch
